implemented paging and sorting repository

+++ src/main/java/org/banta/citronix/dto/farm/FarmDTO.java
@@ -31,7 +31,7 @@ public class FarmDTO {
 
     @NotNull(message = "Farm area is required")
     @Min(value = 2000, message = "Farm area must be at least 0.2 hectare (2000 m²)")
-    private Float area;
+    private Double area;
 
     @NotNull(message = "Date established is required")
     @PastOrPresent(message = "Establishment date cannot be in the future")
